+++ client/python/apache_polaris/cli/options/option_tree.py
@@ -290,12 +290,6 @@ class OptionTree:
                                 Hints.PROPERTY,
                                 allow_repeats=True,
                             ),
-                            Argument(
-                                Arguments.LABEL,
-                                str,
-                                Hints.LABEL,
-                                allow_repeats=True,
-                            ),
                         ]
                         + OptionTree._FEDERATION_ARGS,
                         input_name=Arguments.CATALOG,
@@ -309,13 +303,7 @@ class OptionTree:
                                 Arguments.PRINCIPAL_ROLE,
                                 str,
                                 Hints.PrincipalRoles.PRINCIPAL_ROLE,
-                            ),
-                            Argument(
-                                Arguments.LABEL_FILTER,
-                                str,
-                                Hints.LABEL_FILTER,
-                                allow_repeats=True,
-                            ),
+                            )
                         ],
                     ),
                     Option(
@@ -346,17 +334,6 @@ class OptionTree:
                                 str,
                                 Hints.REMOVE_PROPERTY,
                                 allow_repeats=True,
-                            ),
-                            Argument(
-                                Arguments.LABEL,
-                                str,
-                                Hints.LABEL,
-                                allow_repeats=True,
-                            ),
-                            Argument(
-                                Arguments.CLEAR_LABELS,
-                                bool,
-                                Hints.CLEAR_LABELS,
                             ),
                         ],
                         input_name=Arguments.CATALOG,